Some simple refactoring
